+++ packages/select/src/components/base-select-mobile/Component.tsx
@@ -323,7 +323,15 @@ export const BaseSelectMobile = forwardRef(
 
         const handleFieldKeyDown = (event: KeyboardEvent<HTMLDivElement | HTMLInputElement>) => {
             inputProps.onKeyDown(event);
-            if (autocomplete && !open && (event.key.length === 1 || event.key === 'Backspace')) {
+
+            // https://caniuse.com/?search=KeyboardEvent.key
+            const isKeyUnsupported = event.key === 'Unidentified';
+
+            if (
+                autocomplete &&
+                !open &&
+                (isKeyUnsupported || event.key.length === 1 || event.key === 'Backspace')
+            ) {
                 // Для автокомплита - открываем меню при начале ввода
                 openMenu();
             }
@@ -540,6 +548,7 @@ export const BaseSelectMobile = forwardRef(
                             hasCloser={true}
                             sticky={true}
                             {...modalHeaderProps}
+                            title={undefined}
                             bottomAddons={
                                 <React.Fragment>
                                     {renderSearch()}
@@ -547,7 +556,7 @@ export const BaseSelectMobile = forwardRef(
                                 </React.Fragment>
                             }
                         >
-                            {label || placeholder}
+                            {modalHeaderProps?.title || label || placeholder}
                         </ModalMobile.Header>
 
                         <ModalMobile.Content flex={true} className={styles.modalContent}>